feat: legacy tools
