+++ ee/hogai/graph/root/prompts.py
@@ -139,107 +139,6 @@ Follow these guidelines when creating a dashboard:
 New memories will automatically be added to the core memory as the conversation progresses. If users ask to save, update, or delete the core memory, say you have done it. If the '/remember [information]' command is used, the information gets appended verbatim to core memory.
 """.strip()
 
-SESSION_SUMMARIZATION_PROMPT_BASE = """
-<session_summarization>
-The tool `session_summarization` helps you to summarize session recordings by analysing the events within those sessions.
-
-{{{conditional_context}}}
-
-Synonyms:
-- "summarize": "watch", "analyze", "review", and similar
-- "session recordings": "sessions", "recordings", "replays", "user sessions", and similar
-
-Follow these guidelines when summarizing session recordings:
-- CRITICAL: Always pass the user's complete, unmodified query to the `session_summarization_query` parameter
-- DO NOT truncate, summarize, or extract keywords from the user's query
-- The query is used to find relevant sessions - context helps find better matches
-- Use explicit tool definition to make a decision
-</session_summarization>
-"""
-
-SESSION_SUMMARIZATION_PROMPT_NO_REPLAY_CONTEXT = """
-There are no current filters in the user's UI context. It means that you need to:
-- Convert the user query into a `session_summarization_query`
-- The query should be used to search for relevant sessions and then summarize them
-- Assume the `should_use_current_filters` should be always `false`
-- Generate the `summary_title` based on the user's query
-"""
-
-SESSION_SUMMARIZATION_PROMPT_WITH_REPLAY_CONTEXT = """
-There are current filters in the user's UI context. It means that you need to:
-- Convert the user query into a `session_summarization_query`
-- The query should be used to understand the user's intent
-- Decide if the query is relevant to the current filters and set `should_use_current_filters` accordingly
-- Generate the `summary_title` based on the user's query and the current filters
-
-```json
-{{{current_filters}}}
-```
-"""
-
-
-ROOT_INSIGHT_DESCRIPTION_PROMPT = """
-Pick the most suitable visualization type for the user's question.
-
-## `trends`
-
-A trends insight visualizes events over time using time series. They're useful for finding patterns in historical data.
-
-The trends insights have the following features:
-- The insight can show multiple trends in one request.
-- Custom formulas can calculate derived metrics, like `A/B*100` to calculate a ratio.
-- Filter and break down data using multiple properties.
-- Compare with the previous period and sample data.
-- Apply various aggregation types, like sum, average, etc., and chart types.
-- And more.
-
-Examples of use cases include:
-- How the product's most important metrics change over time.
-- Long-term patterns, or cycles in product's usage.
-- The usage of different features side-by-side.
-- How the properties of events vary using aggregation (sum, average, etc).
-- Users can also visualize the same data points in a variety of ways.
-
-## `funnel`
-
-A funnel insight visualizes a sequence of events that users go through in a product. They use percentages as the primary aggregation type. Funnels use two or more series, so the conversation history should mention at least two events.
-
-The funnel insights have the following features:
-- Various visualization types (steps, time-to-convert, historical trends).
-- Filter data and apply exclusion steps.
-- Break down data using a single property.
-- Specify conversion windows, details of conversion calculation, attribution settings.
-- Sample data.
-- And more.
-
-Examples of use cases include:
-- Conversion rates.
-- Drop off steps.
-- Steps with the highest friction and time to convert.
-- If product changes are improving their funnel over time.
-- Average/median time to convert.
-- Conversion trends over time.
-
-## `retention`
-
-A retention insight visualizes how many users return to the product after performing some action. They're useful for understanding user engagement and retention.
-
-The retention insights have the following features: filter data, sample data, and more.
-
-Examples of use cases include:
-- How many users come back and perform an action after their first visit.
-- How many users come back to perform action X after performing action Y.
-- How often users return to use a specific feature.
-
-## 'sql'
-
-The 'sql' insight type allows you to write arbitrary SQL queries to retrieve data.
-
-The SQL insights have the following features:
-- Filter data using arbitrary SQL.
-- All ClickHouse SQL features.
-- You can nest subqueries as needed.
-""".strip()
 
 ROOT_HARD_LIMIT_REACHED_PROMPT = """
 You have reached the maximum number of iterations, a security measure to prevent infinite loops. Now, summarize the conversation so far and answer my question if you can. Then, ask me if I'd like to continue what you were doing.
@@ -247,7 +146,7 @@ You have reached the maximum number of iterations, a security measure to prevent
 
 ROOT_BILLING_CONTEXT_WITH_ACCESS_PROMPT = """
 <billing_context>
-If the user asks about billing, their subscription, their usage, or their spending, use the `retrieve_billing_information` tool to answer.
+If the user asks about billing, their subscription, their usage, or their spending, use the `ReadData` tool with the `billing_info` kind to answer.
 You can use the information retrieved to check which PostHog products and add-ons the user has activated, how much they are spending, their usage history across all products in the last 30 days, as well as trials, spending limits, billing period, and more.
 If the user wants to reduce their spending, always call this tool to get suggestions on how to do so.
 If an insight shows zero data, it could mean either the query is looking at the wrong data or there was a temporary data collection issue. You can investigate potential dips in usage/captured data using the billing tool.
