+++ ee/hogai/graph/root/tools/__init__.py
@@ -1,3 +1,4 @@
+from .legacy import create_and_query_insight, create_dashboard, session_summarization
 from .navigate import NavigateTool, NavigateToolArgs
 from .read import ReadDataTool, ReadDataToolArgs
 from .read_taxonomy import ReadTaxonomyTool
@@ -14,4 +15,7 @@ __all__ = [
     "TodoWriteToolArgs",
     "NavigateTool",
     "NavigateToolArgs",
+    "create_and_query_insight",
+    "session_summarization",
+    "create_dashboard",
 ]
